+++ np.Templating/CHANGELOG.md
@@ -24,6 +24,9 @@ See Plugin [Documentation](https://noteplan.co/templates/docs) for details on av
 - Add `journalingQuestion` commands to WebModule per Tim Shaker - https://discord.com/channels/763107030223290449/963950027946999828/1051665188648132648
 - Add `date.daysUntil` to DateModule
 - Fix bug in promises in date shorthand codes
+- add note.currentNote() to NoteModule
+- fixed formattedDateTime to work with strftime format (what it was) or moment (what we use everywhere else)
+
 
 ## [1.12.0] 2025-03-09 @dwertheimer